# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shumi API",
        "title": "Shumi API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/me": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Профиль пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.User"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Обновить информацию профиля пользователя",
                "parameters": [
                    {
                        "description": "Имя пользователя",
                        "name": "name",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Локация пользователя",
                        "name": "location",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Пол пользователя",
                        "name": "gender",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Дата рождения пользователя",
                        "name": "birth_date",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Информация о пользователе",
                        "name": "info",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Интересы о пользователе",
                        "name": "interests",
                        "in": "body",
                        "schema": {
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно обновлен"
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
        "/api/me/photo": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Обновить информацию профиля пользователя",
                "parameters": [
                    {
                        "description": "Фото пользователя",
                        "name": "user_photo",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно обновлен"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
        "/api/me/widgets": {
            "put": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Обновить виджет",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "collectionFormat": "csv",
                        "description": "Данные виджета",
                        "name": "data",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "description": "Файлы виджета",
                        "name": "files",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Виджет успешно обновлен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Создать виджет",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "collectionFormat": "csv",
                        "description": "Данные виджета",
                        "name": "data",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "description": "Файлы виджета",
                        "name": "files",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Виджет успешно создан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Удалить виджет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id виджета",
                        "name": "widget_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Виджет удален",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
        "/api/me/widgets/order": {
            "put": {
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Обновить очередь виджета",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id виджета",
                        "name": "widget_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Изменение позиции виджета виджета",
                        "name": "order",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Очередь успешно обновлена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
        "/api/me/widgets/photo": {
            "delete": {
                "summary": "Удалить фото виджета",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id виджета",
                        "name": "widget_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Позиция фото в списке",
                        "name": "index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фото виджета удалено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/database.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "database.Error": {
            "description": "Ошибка",
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "database.User": {
            "description": "Основные данные профиля и связанные виджеты.",
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hide": {
                    "type": "boolean"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "location": {
                    "type": "string"
                },
                "miniapp_name": {
                    "type": "string"
                },
                "tg_user": {
                    "type": "string"
                },
                "user_info": {
                    "type": "string"
                },
                "user_photo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Widget"
                    }
                }
            }
        },
        "database.Vote": {
            "description": "Голос в опросе",
            "type": "object",
            "properties": {
                "option": {
                    "type": "integer"
                },
                "survey": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "database.Widget": {
            "description": "Виджет",
            "type": "object",
            "properties": {
                "additionalData": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/database.User"
                },
                "vote": {
                    "$ref": "#/definitions/database.Vote"
                }
            }
        },
...
    }
}